attack-capital/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
livekit-api>=0.5.0
livekit-agents>=0.1.0
google-generativeai>=0.3.0
faiss-cpu>=1.7.4
numpy==1.26.0
pydantic>=2.4.2
python-multipart==0.0.6
sqlalchemy>=2.0.21
sentence-transformers==2.2.2
websockets>=12.0,<13.0
mem0ai>=0.1.117
requests>=2.30.0
groq==0.31.1
langchain-groq==0.3.7
langchain-core>=0.3.75
groq==0.31.1
langchain-groq==0.3.7
langchain-core>=0.3.75